Updated includes and booleans to latest available
Some of the includes were deprecated, and the boolean "-" was apparently changed to "~" in Python3.  I revised accordingly.  I also dropped a problematic section on resizing the background image due to conflicts with the imageio library.  So the background doesn't get resized, but otherwise it works and looks great.  I'm sure a more professional programmer could resolve that quickly, but I'm just a hobbyist.  Thanks to the authors for their hard work on this cool program.